+++ packages/epoint/src/integrify/epoint/schemas/request.py
@@ -1,8 +1,9 @@
 from decimal import Decimal
+from typing import Any
 
 from integrify.epoint import env
 from integrify.schemas import PayloadBaseModel
-from pydantic import Field
+from pydantic import AliasChoices, Field
 
 
 class MinimalPaymentRequestSchema(PayloadBaseModel):
@@ -66,3 +67,40 @@ class SplitPayAndSaveCardRequestSchema(BasePaymentRequestSchema):
     split_user: str = Field(validation_alias='split_user_id')
     split_amount: Decimal
     description: str | None = None
+
+
+##############################################################################
+# Apple Pay & Google Pay
+class CreateWidgetRequestSchema(PayloadBaseModel):
+    amount: Decimal
+    order_id: str
+    description: str
+
+
+class CreateTokenPaymentRequestSchema(PayloadBaseModel):
+    amount: Decimal
+    currency: str
+    order_id: str
+    description: str | None = None
+
+
+class ApplePaySessionRequestSchema(PayloadBaseModel):
+    origin: str
+
+
+class BaseTokenPayRequestSchema(PayloadBaseModel):
+    id: int | str = Field(validation_alias=AliasChoices('payment_id', 'id'))
+    token: Any
+    billing_contact: Any = Field(
+        default=None,
+        validation_alias=AliasChoices('billing_contact', 'billingContact'),
+        serialization_alias='billingContact',
+    )
+
+
+class ApplePayRequestSchema(BaseTokenPayRequestSchema):
+    pass
+
+
+class GooglePayRequestSchema(BaseTokenPayRequestSchema):
+    pass